refactor(bpa-server): flatten module structure, move build to main, split config
