--- bpa-server/src/bpa/policy.rs
@@ -1,27 +1,0 @@
-use super::*;
-use serde::{Deserialize, Serialize};
-
-// The EgressPolicy enum separates the type from its configuration.
-#[derive(Serialize, Deserialize, Debug)]
-#[serde(tag = "type", content = "config")]
-pub enum EgressPolicyConfig {
-    // #[cfg(feature = "htb_policy")]
-    // Htb { config: hardy_bpa::policy::htb_policy::HtbConfig },
-
-    // #[cfg(feature = "tbf_policy")]
-    // Tbf { config: hardy_bpa::policy::tbf_policy::TbfConfig },
-
-    // Catch-all for unrecognised policy types (logged as a warning).
-    #[serde(other)]
-    Unknown,
-}
-
-pub fn new(config: EgressPolicyConfig) -> anyhow::Result<Arc<dyn hardy_bpa::policy::EgressPolicy>> {
-    match config {
-        // #[cfg(feature = "htb_policy")]
-        // EgressPolicyConfig::Htb { config } => {
-        //     Ok(Arc::new(hardy_bpa::policy::htb_policy::HtbPolicy::new(config)))
-        // }
-        EgressPolicyConfig::Unknown => Err(anyhow::anyhow!("Unknown policy type")),
-    }
-}
